# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HT Lotto API",
        "description": "REST API for HT ភ្នាក់ Lottery Agent App. Roles: admin | master | staff.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/report/summary": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Summary report — admin: all staff; master: own staff only",
                "operationId": "e7578ce2b43eb54378f547bd146d5e87",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-05-01"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-05-02"
                        }
                    },
                    {
                        "name": "staff_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "session",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "morning",
                                "noon",
                                "evening"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Summary",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ReportSummary"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Staff forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/report/daily": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Daily sales breakdown — admin: all; master: own staff",
                "operationId": "b5744e20579ad0cbbc8896d7b558e9c3",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daily breakdown",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ReportDaily"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Staff forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/report/staff": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Per-staff performance — admin: all; master: own staff",
                "operationId": "e3a18747d6e727e49b4ed085d155ad12",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Staff breakdown",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ReportStaff"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Staff forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/report/export": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Export report data (summary JSON for mobile PDF render) — admin and master only",
                "operationId": "ae82e72756305edd88856125ac43b03b",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Export data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ReportSummary"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "ReportSummary": {
                "properties": {
                    "total_bets": {
                        "type": "integer",
                        "example": 120
                    },
                    "total_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 450000
                    },
                    "total_win": {
                        "type": "number",
                        "format": "float",
                        "example": 80000
                    },
                    "net": {
                        "type": "number",
                        "format": "float",
                        "example": 370000
                    },
                    "by_session": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ReportDaily": {
                "properties": {
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-05-01"
                    },
                    "bets": {
                        "type": "integer",
                        "example": 40
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 150000
                    },
                    "win": {
                        "type": "number",
                        "format": "float",
                        "example": 20000
                    }
                },
                "type": "object"
            },
            "ReportStaff": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 3
                    },
                    "name": {
                        "type": "string",
                        "example": "បង ម៉េន"
                    },
                    "bets": {
                        "type": "integer",
                        "example": 40
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 150000
                    },
                    "win": {
                        "type": "number",
                        "format": "float",
                        "example": 20000
                    },
                    "net": {
                        "type": "number",
                        "format": "float",
                        "example": 130000
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}